Less dependencies in generated C++ headers
